Add GpioInputMatrix, implement MillisecondTimer
- Add class GpioInputMatrix that can read elements in i.e. a button
matrix, and debounce those read states.

- Implement the MillisecondTimer interface with a class Microsecond
PeriodMillisecondTimer that allows for tracking milliseconds while
still triggering interrupts in microsecond intervals.

- Add new startup-file for concentrating only on the led matrix and
button matrix, and rewrite main to use the matrices.

- Later on a shift register chip can be used to expand the number of
GPIOs, so that both step leds/buttons can be used with the rotary
encoder and display, but for now focus is shifted to programming
rythms and not on controlling tempo