Port CallbackScheduler to use r2k::function
Change CallbackScheduler to use r2k::function as the underlying type
for storing functions to call in the schedule, so that any callable
type can be used together with the scheduler.

Update main to use a closure where the number of the LED to turn off
is captured when making the scheduling.